requirements.txt
streamlit==1.29.0
spacy==3.7.2
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.6.1
pandas==2.1.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
